# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eventos API",
        "description": "Descripción de tu API",
        "contact": {
            "email": "roni290690@email.com"
        },
        "license": {
            "name": "Licencia",
            "url": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/eventos/search-participante/{codigo}/{documento}": {
            "get": {
                "tags": [
                    "Participantes"
                ],
                "summary": "Listar Participante por DNI o Codigo",
                "operationId": "searchParticipante",
                "parameters": [
                    {
                        "name": "codigo",
                        "in": "path",
                        "description": "Codigo del Participante",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "documento",
                        "in": "path",
                        "description": "Número Documento del Participante",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/sortear-participantes": {
            "get": {
                "tags": [
                    "Participantes"
                ],
                "summary": "Sortear Participante",
                "description": "Reguest",
                "operationId": "sortearParticipante",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/lista-participantes/{elementosPorPagina}": {
            "get": {
                "tags": [
                    "Participantes"
                ],
                "summary": "Listar Participantes",
                "description": "Reguest",
                "operationId": "getParticipante",
                "parameters": [
                    {
                        "name": "elementosPorPagina",
                        "in": "path",
                        "description": "Elementos por Pagina",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/lista-premio": {
            "get": {
                "tags": [
                    "Premios"
                ],
                "summary": "Listar Premios",
                "description": "Reguest",
                "operationId": "listarPremio",
                "responses": {
                    "200": {
                        "description": "Respuesta exitosa"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/lista-premio/{idPremio}": {
            "get": {
                "tags": [
                    "Premios"
                ],
                "summary": "Listar Premio por ID",
                "operationId": "showPremioId",
                "parameters": [
                    {
                        "name": "idPremio",
                        "in": "path",
                        "description": "ID del Premio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/create-premio": {
            "post": {
                "tags": [
                    "Premios"
                ],
                "summary": "Crear Premio",
                "operationId": "storePremio",
                "requestBody": {
                    "description": "Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Premios"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/update-premio/{idPremio}": {
            "put": {
                "tags": [
                    "Premios"
                ],
                "summary": "Actualizar Premio",
                "operationId": "updatePremio",
                "parameters": [
                    {
                        "name": "idPremio",
                        "in": "path",
                        "description": "ID del Premio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Premios"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "datos actualizado correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/eventos/eliminar-premio/{idPremio}": {
            "delete": {
                "tags": [
                    "Premios"
                ],
                "summary": "Eliminar Juego",
                "operationId": "destroyPremio",
                "parameters": [
                    {
                        "name": "idPremio",
                        "in": "path",
                        "description": "ID del Premio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "datos Elimnados Correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Premios": {
                "required": [
                    "nombre",
                    "idUsuario"
                ],
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "idUsuario": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}